agent/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
stagehand>=0.3.0
httpx>=0.24.0
pydantic>=2.0.0
modal>=0.73.0
fastapi[standard]>=0.115.0
# LangGraph agent dependencies (using Gemini)
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
google-generativeai>=0.8.0

